+++ packages/component/src/block/orderEntry/index.tsx
@@ -24,7 +24,7 @@ import {
   useDebounce,
   useMediaQuery,
 } from "@orderly.network/hooks";
-import type { UseOrderEntryMetaState } from "@orderly.network/hooks";
+import { UseOrderEntryMetaState, utils } from "@orderly.network/hooks";
 
 import {
   API,
@@ -140,7 +140,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
       true
     );
 
-    const baseDP = symbolConfig?.base_dp
+    const baseDP = symbolConfig?.base_dp;
 
     const ee = useEventEmitter();
     const isMarketOrder = [OrderType.MARKET, OrderType.STOP_MARKET].includes(
@@ -226,6 +226,17 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
 
     // const isTable = useMediaQuery(MEDIA_TABLET);
 
+    const formatQty = () => {
+      // const dp = new Decimal(symbolConfig?.base_tick || "0").toNumber();
+      // TODO: optimization
+      // if (dp < 1) return;
+      const quantity = utils.formatNumber(
+        formattedOrder?.order_quantity,
+        new Decimal(symbolConfig?.base_tick || "0").toNumber()
+      );
+      props.onFieldChange("order_quantity", quantity);
+    }
+
     const onFocus = (type: InputType) => (_: FocusEvent<HTMLInputElement>) => {
       currentFocusInput.current = type;
     };
@@ -235,6 +246,10 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
         if (currentFocusInput.current !== type) return;
         currentFocusInput.current = InputType.NONE;
       }, 300);
+
+      if (type === InputType.QUANTITY) {
+        formatQty();
+      }
     };
 
     const onSubmit = (event: FormEvent) => {
@@ -256,10 +271,13 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
             setErrorsVisible(true);
             return Promise.reject("cancel");
           }
+          formatQty();
           if (needConfirm) {
             return modal.confirm({
               maxWidth: "sm",
               title: "Confirm Order",
+              okId: "orderly-confirm-order-dialog-confirm",
+              cancelId: "orderly-confirm-order-dialog-cancel",
               onCancel: () => {
                 return Promise.reject("cancel");
               },
@@ -386,17 +404,20 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
       >
         <div className="orderly-flex orderly-flex-col orderly-gap-3 orderly-text-3xs">
           <SegmentedButton
+            id="orderly-order-entry-trade-type"
             buttons={[
               {
+                id: "orderly-order-entry-buy-button",
                 label: "Buy",
                 value: OrderSide.BUY,
                 disabled,
                 activeClassName:
-                  "orderly-bg-trade-profit orderly-text-base-contrast after:orderly-bg-trade-profit orderly-font-bold desktop:orderly-font-bold",
+                "orderly-bg-trade-profit orderly-text-base-contrast after:orderly-bg-trade-profit orderly-font-bold desktop:orderly-font-bold",
                 disabledClassName:
-                  "orderly-bg-base-400 orderly-text-base-contrast-20 after:orderly-bg-base-400 orderly-cursor-not-allowed orderly-font-bold desktop:orderly-font-bold",
+                "orderly-bg-base-400 orderly-text-base-contrast-20 after:orderly-bg-base-400 orderly-cursor-not-allowed orderly-font-bold desktop:orderly-font-bold",
               },
               {
+                id: "orderly-order-entry-sell-button",
                 label: "Sell",
                 value: OrderSide.SELL,
                 disabled,
@@ -438,16 +459,17 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
             </Button>
           </div>
           <MSelect
+            id="orderly-order-entry-order-type"
             label={"Order Type"}
             value={formattedOrder.order_type}
             className="orderly-bg-base-600 orderly-font-semibold"
             color={side === OrderSide.BUY ? "buy" : "sell"}
             fullWidth
             options={[
-              { label: "Limit order", value: "LIMIT" },
-              { label: "Market order", value: "MARKET" },
-              { label: "Stop limit", value: "STOP_LIMIT" },
-              { label: "Stop market", value: "STOP_MARKET" },
+              { label: "Limit order", value: "LIMIT", id: "orderly-order-entry-order-type-limit" },
+              { label: "Market order", value: "MARKET", id: "orderly-order-entry-order-type-market" },
+              { label: "Stop limit", value: "STOP_LIMIT", id: "orderly-order-entry-order-type-stop-limit" },
+              { label: "Stop market", value: "STOP_MARKET", id: "orderly-order-entry-order-type-stop-market" },
             ]}
             onChange={(value) => {
               // field.onChange(value);
@@ -571,13 +593,14 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
               %
             </span>
             <button
+              id="orderly-order-entry-max-buy-or-sell"
               type="button"
               className="orderly-flex orderly-items-center orderly-gap-1 orderly-tabular-nums"
               onClick={() => {
                 props.onFieldChange("order_quantity", maxQty);
               }}
             >
-              <span className="orderly-text-base-contrast-54">
+              <span id="orderly-order-entry-max-buy-or-sell-text" className="orderly-text-base-contrast-54">
                 {formattedOrder.side === OrderSide.BUY ? "Max buy" : "Max sell"}
               </span>
               <Numeral precision={baseDP}>{maxQty}</Numeral>
@@ -633,7 +656,7 @@ export const OrderEntry = forwardRef<OrderEntryRef, OrderEntryProps>(
             reduceOnly={formattedOrder.reduce_only}
             onFieldChange={props.onFieldChange}
           />
-          <StatusGuardButton>
+          <StatusGuardButton id="orderly-order-entry-status-guard-button">
             <Button
               id="orderly-order-entry-confirm-button"
               className="orderly-text-xs desktop:orderly-font-bold desktop:orderly-text-sm"