+++ packages/component/src/block/walletConnect/walletConnect.tsx
@@ -4,8 +4,12 @@ import { Switch } from "@/switch";
 import { FC, useCallback, useContext, useMemo, useState } from "react";
 import { AccountStatusEnum } from "@orderly.network/types";
 import { StepItem } from "./sections/step";
-import { OrderlyContext, useAccount, useMutation } from "@orderly.network/hooks";
-
+import {
+  useAccount,
+  useCheckReferralCode,
+  useGetReferralCode,
+  useMutation,
+} from "@orderly.network/hooks";
 import Button from "@/button";
 import { toast } from "@/toast";
 import { RememberMe } from "./sections/rememberMe";
@@ -25,6 +29,7 @@ export interface WalletConnectProps {
 export const WalletConnect: FC<WalletConnectProps> = (props) => {
   // const { status = AccountStatusEnum.NotConnected } = props;
   const {
+    account,
     state: { status, isNew },
   } = useAccount();
 
@@ -32,13 +37,13 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
   const [remember, setRemember] = useState(true);
   const localRefCode = localStorage.getItem("referral_code") || undefined;
   const [refCode, setRefCode] = useState<string | undefined>(localRefCode);
-  const [
-    bindRefCode,
-    { error: updateOrderError, isMutating: updateMutating },
-  ] = useMutation("/v1/referral/bind", "POST");
-
+  const [bindRefCode, { error: updateOrderError, isMutating: updateMutating }] =
+    useMutation("/v1/referral/bind", "POST");
 
   const { referral } = useContext(OrderlyAppContext);
+  const { isExist } = useCheckReferralCode(refCode);
+
+  const { referral_code, isLoading: loadingReferralCode } = useGetReferralCode(account?.accountId);
 
   const buttonLabel = useMemo(() => {
     if (status < AccountStatusEnum.SignedIn) {
@@ -50,45 +55,54 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
     return "--";
   }, [status]);
 
+  const enableTrading = useCallback(() => {
+    setHandleStep(2);
+    return props
+      .onEnableTrading?.(remember)
+      .then(
+        () => {
+          if (refCode && refCode.length > 0) {
+            bindRefCode({ referral_code: refCode })
+              .then((res) => {
+                referral?.onBoundRefCode?.(true, undefined);
+              })
+              .catch((e) => {
+                referral?.onBoundRefCode?.(false, e);
+              })
+              .finally(() => {
+                localStorage.removeItem("referral_code");
+              });
+          }
+          props.onComplete?.();
+        },
+        (e) => {
+          let errorText = `${e}`;
+          if ("message" in e) {
+            errorText = e.message;
+          }
+
+          if ("referral code not exist" === errorText) {
+            errorText = "This referral code does not exist";
+          }
+          toast.error(errorText);
+        }
+      )
+      .finally(() => {
+        setHandleStep(0);
+      });
+  }, [refCode]);
+
   const onClick = useCallback(() => {
     if (status < AccountStatusEnum.SignedIn) {
       setHandleStep(1);
       return props.onSignIn?.().finally(() => {
         setHandleStep(0);
+        // sign in success then auto enable trading
+        enableTrading();
       });
     }
     if (status < AccountStatusEnum.EnableTrading) {
-      setHandleStep(2);
-      return props
-        .onEnableTrading?.(remember)
-        .then(
-          () => {
-            if (refCode && refCode.length > 0) {
-              bindRefCode({ referral_code: refCode }).then((res) => {
-                referral?.onBoundRefCode?.(true, undefined);
-              }).catch((e) => {
-                referral?.onBoundRefCode?.(false, e);
-               }).finally(() => {
-                localStorage.removeItem("referral_code");
-              });
-            }
-            props.onComplete?.();
-          },
-          (e) => {
-            let errorText = `${e}`;
-            if ("message" in (e)) {
-              errorText = e.message;
-            }
-
-            if ("referral code not exist" === errorText) {
-              errorText = "This referral code does not exist";
-            }
-            toast.error(errorText);
-          }
-        )
-        .finally(() => {
-          setHandleStep(0);
-        });
+      return enableTrading();
     }
   }, [status, remember, refCode]);
 
@@ -140,7 +154,15 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
         />
       </Paper>
 
-      <ReferralCode className="orderly-pt-5" refCode={refCode} setRefCode={setRefCode} />
+      {(referral_code?.length || 0) === 0 && loadingReferralCode === false && (
+        <ReferralCode
+          className="orderly-pt-5"
+          refCode={refCode}
+          setRefCode={setRefCode}
+          isExist={isExist}
+        />
+      )}
+
       <div className="orderly-pt-5 orderly-pb-7 orderly-flex orderly-justify-between orderly-items-center">
         <RememberMe />
         <div>
@@ -153,9 +175,10 @@ export const WalletConnect: FC<WalletConnectProps> = (props) => {
       </div>
       <div>
         <Button
+          id="orderly-wallet-connector-button"
           className="orderly-text-xs orderly-text-base-contrast"
           fullWidth
-          disabled={handleStep > 0}
+          disabled={handleStep > 0 || (isExist !== true && (refCode?.length || 0) > 0)}
           onClick={onClick}
           loading={handleStep > 0}
         >