+++ packages/component/src/index.ts
@@ -132,3 +132,5 @@ export type { OrderlyConfig, OrderlyConfigCtx } from "./types/orderly.config";
 export { installExtension, ExtensionPosition } from "./plugin";
 
 export * from "./datePicker";
+
+export * from "./switch";