+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @orderly.network/futures
 
+## 3.3.2
+
+### Patch Changes
+
+- f7bfae80: fix bug;
+- Updated dependencies [f7bfae80]
+  - @orderly.network/types@1.5.2
+  - @orderly.network/utils@1.5.2
+
+## 3.3.2-alpha.0
+
+### Patch Changes
+
+- fix bug;
+- Updated dependencies
+  - @orderly.network/types@1.5.2-alpha.0
+  - @orderly.network/utils@1.5.2-alpha.0
+
+## 3.3.1
+
+### Patch Changes
+
+- 34efc6c0: update
+- Updated dependencies [34efc6c0]
+  - @orderly.network/types@1.5.1
+  - @orderly.network/utils@1.5.1
+
+## 3.3.1-alpha.0
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@1.5.1-alpha.0
+  - @orderly.network/utils@1.5.1-alpha.0
+
 ## 3.3.0
 
 ### Minor Changes