+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,29 @@
 # tsconfig
 
+## 0.3.2
+
+### Patch Changes
+
+- f7bfae80: fix bug;
+
+## 0.3.2-alpha.0
+
+### Patch Changes
+
+- fix bug;
+
+## 0.3.1
+
+### Patch Changes
+
+- 34efc6c0: update
+
+## 0.3.1-alpha.0
+
+### Patch Changes
+
+- update
+
 ## 0.3.0
 
 ### Minor Changes